+++ app/domain/infra/gemini/gemini.go
@@ -1,0 +1,7 @@
+package gemini
+
+import "context"
+
+type IClient interface {
+	GenAnswer(ctx context.Context, q string) (string, error)
+}
